build deep q-learning to get better strategy
